# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API sobre el Sistema de Transporte Colectivo (STC) de la Ciudad de México",
        "title": "Apimetro",
        "contact": {
            "name": "galigaribaldi (Galileo Cabrera Garibaldi)",
            "email": "galigaribaldi0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/movilidad",
    "paths": {
...
        "/mapas/geojsonLinea": {
            "get": {
                "description": "Obtener líneas de transporte en formato GeoJSON con métricas operativas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeoJSON"
                ],
                "summary": "GeoJSON Lineas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by sistema",
                        "name": "sistema",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by num_comercial",
                        "name": "num_comercial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by nombre_ramal",
                        "name": "nombre_ramal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by jerarquia_transporte",
                        "name": "jerarquia_transporte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by derecho_de_via",
                        "name": "derecho_de_via",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by es_cetram",
                        "name": "es_cetram",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sentido",
                        "name": "sentido",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by cetram_real (250m radius)",
                        "name": "cetram_real",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by existe (true/false)",
                        "name": "existe",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FeatureCollection"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/{sistema}/linea": {
            "get": {
                "description": "Obtener datos de líneas de transporte con filtros avanzados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linea"
                ],
                "summary": "Datos de Linea",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID interno de línea",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre de la línea",
                        "name": "nombre",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Número comercial",
                        "name": "num_comercial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre del ramal",
                        "name": "nombre_ramal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Clasificación operativa",
                        "name": "clasificacion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Longitud en kilómetros",
                        "name": "tam_km",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar líneas operativas",
                        "name": "existe",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Linea"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Crear una nueva línea de transporte",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linea"
                ],
                "summary": "Crear Linea",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la línea a crear",
                        "name": "linea",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Linea"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Linea"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Eliminar una línea de transporte por su ID (incluye eliminación en cascada de ramales y descripciones)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linea"
                ],
                "summary": "Eliminar Linea",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la línea a eliminar",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualizar los datos de una línea de transporte por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linea"
                ],
                "summary": "Actualizar Linea",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la línea a actualizar",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la línea",
                        "name": "linea",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
        "models.DescripcionLinea": {
            "type": "object",
            "properties": {
                "anio_fin_ampliacion": {
                    "type": "integer"
                },
                "anio_inicio_ampliacion": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "descripcion_linea_id": {
                    "type": "integer"
                },
                "direccion": {
                    "type": "string"
                },
                "inicio_original": {
                    "type": "string"
                },
                "linea_base": {
                    "type": "integer"
                },
                "terminal_original": {
                    "type": "string"
                },
                "tipo_linea": {
                    "type": "string"
                }
            }
        },
...
        "models.Feature": {
            "type": "object",
            "properties": {
                "geometry": {
                    "description": "Geometría de tipo Point",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Geometry"
                        }
                    ]
                },
                "properties": {
                    "description": "Atributos dinámicos",
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "description": "\"Feature\"",
                    "type": "string",
                    "example": "Feature"
                }
            }
        },
        "models.FeatureCollection": {
            "type": "object",
            "properties": {
                "features": {
                    "description": "Lista de features (líneas o estaciones)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Feature"
                    }
                },
                "type": {
                    "description": "\"FeatureCollection\"",
                    "type": "string",
                    "example": "FeatureCollection"
                }
            }
        },
        "models.Geometry": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "description": "interface{} permite múltiples niveles de arreglos"
                },
                "type": {
                    "description": "\"Point\"",
                    "type": "string",
                    "example": "Point"
                }
            }
        },
        "models.Linea": {
            "type": "object",
            "properties": {
                "anio_inauguracion": {
                    "type": "integer",
                    "example": 1969
                },
                "capacidad_vehiculo": {
                    "type": "integer",
                    "example": 1000
                },
                "clasificacion": {
                    "type": "string",
                    "example": "existente"
                },
                "color_en": {
                    "type": "string",
                    "example": "Pink"
                },
                "color_esp": {
                    "type": "string",
                    "example": "Rosa"
                },
                "derecho_de_via": {
                    "type": "string",
                    "example": "Superficie"
                },
                "descripcion_linea": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DescripcionLinea"
                    }
                },
                "existe": {
                    "type": "boolean",
                    "example": true
                },
                "geom": {
                    "type": "string",
                    "example": "MULTILINESTRING(...)"
                },
                "jerarquia_transporte": {
                    "type": "string",
                    "example": "Línea principal"
                },
                "linea_base_ramal": {
                    "type": "integer",
                    "example": 1
                },
                "linea_id": {
                    "type": "integer",
                    "example": 1
                },
                "nombre": {
                    "type": "string",
                    "example": "Línea 1"
                },
                "num_comercial": {
                    "type": "string",
                    "example": "1"
                },
                "ramal_id": {
                    "type": "integer",
                    "example": 1
                },
                "ramales": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ramal"
                    }
                },
                "route_gtfs": {
                    "type": "string",
                    "example": "L1"
                },
                "sistema": {
                    "type": "string",
                    "example": "METRO"
                },
                "tam_km": {
                    "type": "number",
                    "example": 18.8
                }
            }
        },
        "models.Ramal": {
            "type": "object",
            "properties": {
                "anio_creacion": {
                    "type": "integer"
                },
                "estado": {
                    "type": "string"
                },
                "geom": {
                    "type": "string"
                },
                "linea_id": {
                    "type": "integer"
                },
                "nombre_ramal": {
                    "type": "string"
                },
                "ramal_id": {
                    "type": "integer"
                },
                "ramal_num": {
                    "type": "integer"
                },
                "shape_gtfs": {
                    "type": "string"
                },
                "tam_km": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}